scripts/static_data_ingestion_requirements.txt
google-cloud-bigquery==2.34.0
pyarrow==7.0.0
pandas==1.3.5
google-auth==2.6.3